Worked on tests

+++ libmodi/libmodi_handle.c
@@ -1315,7 +1315,7 @@ int libmodi_handle_open_file_io_handle(
 	}
 #endif
 	internal_handle->file_io_handle                   = file_io_handle;
-	internal_handle->file_io_handle_opened_in_library = file_io_handle_opened_in_library;
+	internal_handle->file_io_handle_opened_in_library = (uint8_t) file_io_handle_opened_in_library;
 	internal_handle->access_flags                     = access_flags;
 
 #if defined( HAVE_LIBMODI_MULTI_THREAD_SUPPORT )
